Move theme.ts and types.ts under src/arches
