+++ arches/app/src/controlledlists/components/editor/AddValue.vue
@@ -5,7 +5,7 @@ import { useGettext } from "vue3-gettext";
 
 import Button from "primevue/button";
 
-import { ARCHES_CHROME_BLUE } from "@/theme.ts";
+import { ARCHES_CHROME_BLUE } from "@/arches/theme.ts";
 import {
     itemKey,
     ALT_LABEL,
@@ -14,7 +14,7 @@ import {
 } from "@/controlledlists/constants.ts";
 
 import type { Ref } from "vue";
-import type { Language } from "@/types";
+import type { Language } from "@/arches/types";
 import type {
     ControlledListItem,
     Value,
