+++ arches/app/src/controlledlists/components/tree/MoveRow.vue
@@ -25,7 +25,7 @@ import {
 import type { Ref } from "vue";
 import type { TreeExpandedKeys, TreeSelectionKeys } from "primevue/tree/Tree";
 import type { TreeNode } from "primevue/treenode";
-import type { Language } from "@/types";
+import type { Language } from "@/arches/types";
 import type {
     ControlledList,
     ControlledListItem,
