+++ arches/app/src/controlledlists/components/tree/PresentationControls.vue
@@ -9,7 +9,7 @@ import Dropdown from "primevue/dropdown";
 import { selectedLanguageKey } from "@/controlledlists/constants.ts";
 
 import type { Ref } from "vue";
-import type { Language } from "@/types";
+import type { Language } from "@/arches/types";
 
 const { $gettext } = useGettext();
